Get the Windows half building and runnable on Windows (M1)
Capture/Overlay/App were written on Linux and had never compiled on Windows. Fix 27 Vortice 3.6.2 interop mismatches (CreateTextFormat fontCollection arg; out-param CreateTargetForHwnd/CreateSurface; generic BeginDraw<T>; uint D3D11 args on Dispatch/CreateBuffer/EnumAdapters1; Box from Vortice.Mathematics; Matrix3x2 from System.Numerics; WindowFinder's discard was writing to the lambda's nint param) and the WPF/WinForms Application/MessageBox ambiguity. The full solution now builds with 0 errors; the portable half stays green (230 xunit).

Add an --m1 diagnostic mode (RunM1DiagnosticAsync + OverlayRenderer.DrawDiagnosticFrame): a click-through rectangle over the game window with SetWinEventHook tracking, no OCR/DB/pipeline -- the M1 milestone (TZ section 10). Verified on real hardware: the DirectComposition/D2D/DWrite device stack comes up, ex-style confirms click-through/topmost, ~60 fps. Interactive runs get a tray icon so the overlay can be quit; --frames N runs it headless.

+++ src/CznTranslator.Capture/DxgiAdapterProvider.cs
@@ -17,7 +17,7 @@ public sealed class DxgiAdapterProvider : IGraphicsAdapterProvider
 
         for (var index = 0; ; index++)
         {
-            if (factory.EnumAdapters1(index, out var adapter).Failure)
+            if (factory.EnumAdapters1((uint)index, out var adapter).Failure)
                 break;
 
             using (adapter)
